src/core/requirements.txt
# Core FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
 
# Database
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
# If psycopg2-binary fails, alternative
psycopg2>=2.9.0
asyncpg>=0.27.0
 
# Environment and configuration
python-dotenv>=1.0.0
 
# Security and Auth
python-jose[cryptography]>=3.3.0
webauthn>=1.11.0
 
# AI and OpenAI
openai>=1.3.0
tiktoken>=0.5.0
 
# AutoGen packages
autogen-core>=0.4.0
autogen-ext[openai]>=0.4.0
autogen-ext[azure]>=0.4.0
autogen-agentchat>=0.4.0
 
# AutoGen - install specific versions that work
pyautogen>=0.2.0
 
# Data processing
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
 
# Web scraping (optional)
selenium>=4.15.0
webdriver-manager>=4.0.0
 
# Utilities
rich>=13.0.0
python-dateutil>=2.8.0
aiohttp>=3.8.0